Refactor: ModifyPost.tsx 제거 및 구조가 유사한 Write.tsx로 재사용하도록 로직 수정
